module/vision/VisualMesh/VisualMesh/requirements.txt
matplotlib
numpy
opencv-python
pyyaml
tensorflow
tqdm
